+++ tests/ExampleTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Busayo\Mention\Test;
+namespace Unicodeveloper\Mention\Test;
 
 class ExampleTest extends \PHPUnit_Framework_TestCase
 {
-
 
     /**
      * Test that true does in fact equal true